Updated utf_length_n() to count NUL bytes as normal characters.
Previously utf_length_n() would stop at the first NUL byte, even if ‘len’
said that more data was available.  This conforms with the semantics of,
for example, strnlen(), but as we in Ruby can easily get Strings that
contain any byte, we can’t treat NUL bytes separately in the case where
‘len’ is provided (as opposed to utf_length() that is).  So all *_n()
functions will have to be rewritten to handle NUL bytes like any other
byte.  We begin with the utf_length_n() function, because it is an obvious
case where this was causing a problem.

+++ ext/encoding/character/utf-8/unicode.h
@@ -186,7 +186,7 @@ int utf_rindex(const char *haystack, const char *needle);
 int utf_rindex_n(const char *haystack, const char *needle, size_t len);
 bool utf_has_prefix(const char *str, const char *prefix);
 long utf_length(const char *str);
-long utf_length_n(const char *str, size_t len);
+long utf_length_n(const char *str, long len);
 size_t utf_width(const char *str);
 size_t utf_width_n(const char *str, size_t len);
 size_t utf_byte_length(const char *str);
